+++ netlink-packet/src/sock_diag/unix/raw.rs
@@ -1,4 +1,3 @@
-//! for UNIX domain sockets
 #![allow(non_camel_case_types)]
 
 use crate::sock_diag::TcpState::*;
